refactor(stream): refactor append only top n to use one state table
