+++ src/frontend/src/stream_fragmenter/mod.rs
@@ -321,7 +321,7 @@ impl StreamFragmenter {
                 if let Some(table) = &mut top_n_node.table {
                     table.id = state.gen_table_id();
                 } else {
-                    panic!("TopNNode's table shouldn't be None");
+                    panic!("TopN node's table shouldn't be None");
                 }
             }
 
@@ -329,13 +329,16 @@ impl StreamFragmenter {
                 if let Some(table) = &mut group_top_n_node.table {
                     table.id = state.gen_table_id();
                 } else {
-                    panic!("GroupTopNNode's table shouldn't be None");
+                    panic!("Group topN node's table shouldn't be None");
                 }
             }
 
             NodeBody::AppendOnlyTopN(append_only_top_n_node) => {
-                append_only_top_n_node.table_id_l = state.gen_table_id();
-                append_only_top_n_node.table_id_h = state.gen_table_id();
+                if let Some(table) = &mut append_only_top_n_node.table {
+                    table.id = state.gen_table_id();
+                } else {
+                    panic!("Append only TopN node's table shouldn't be None");
+                }
             }
 
             NodeBody::DynamicFilter(dynamic_filter_node) => {
@@ -563,18 +566,24 @@ mod tests {
         {
             // test AppendOnlyTopN Type
             let mut stream_node = StreamNode {
-                node_body: Some(NodeBody::AppendOnlyTopN(AppendOnlyTopNNode {
+                node_body: Some(NodeBody::AppendOnlyTopN(TopNNode {
+                    table: Some(Table {
+                        id: 0,
+                        ..Default::default()
+                    }),
                     ..Default::default()
                 })),
                 ..Default::default()
             };
             StreamFragmenter::assign_local_table_id_to_stream_node(&mut state, &mut stream_node);
-
             if let NodeBody::AppendOnlyTopN(append_only_top_n_node) =
                 stream_node.node_body.as_ref().unwrap()
             {
-                expect_table_id += 2;
-                assert_eq!(expect_table_id, append_only_top_n_node.table_id_h);
+                expect_table_id += 1;
+                assert_eq!(
+                    expect_table_id,
+                    append_only_top_n_node.table.as_ref().unwrap().id
+                );
             }
         }
     }
