+++ src/meta/src/stream/stream_graph.rs
@@ -589,12 +589,9 @@ impl StreamGraphBuilder {
                     }
 
                     NodeBody::AppendOnlyTopN(node) => {
-                        node.table_id_l += table_id_offset;
-                        node.table_id_h += table_id_offset;
-
-                        // TODO add catalog::Table to AppendOnlyTopN
-                        check_and_fill_internal_table(node.table_id_l, None);
-                        check_and_fill_internal_table(node.table_id_h, None);
+                        if let Some(table) = &mut node.table {
+                            update_table(table, "AppendOnlyTopNNode");
+                        }
                     }
                     NodeBody::TopN(node) => {
                         if let Some(table) = &mut node.table {
@@ -1058,8 +1055,9 @@ impl ActorGraphBuilder {
                     .push(node.right_table.as_ref().unwrap().id);
             }
             NodeBody::AppendOnlyTopN(node) => {
-                fragment.state_table_ids.push(node.table_id_l);
-                fragment.state_table_ids.push(node.table_id_h);
+                fragment
+                    .state_table_ids
+                    .push(node.table.as_ref().unwrap().id);
             }
             NodeBody::GroupTopN(node) => {
                 fragment
