+++ src/stream/src/executor/managed_state/top_n/mod.rs
@@ -12,15 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod top_n_bottom_n_state;
 mod top_n_state;
-mod top_n_state_new;
 
-pub use top_n_bottom_n_state::ManagedTopNBottomNState;
-pub use top_n_state::ManagedTopNState;
-pub use top_n_state_new::{ManagedTopNStateNew, TopNStateRow};
-
-pub mod variants {
-    pub const TOP_N_MIN: usize = 0;
-    pub const TOP_N_MAX: usize = 1;
-}
+pub use top_n_state::{ManagedTopNStateNew, TopNStateRow};
