+++ src/stream/src/executor/top_n_appendonly.rs
@@ -14,14 +14,15 @@
 
 use async_trait::async_trait;
 use risingwave_common::array::{Op, StreamChunk};
-use risingwave_common::catalog::{Schema, TableId};
+use risingwave_common::catalog::Schema;
 use risingwave_common::util::ordered::{OrderedRow, OrderedRowDeserializer};
 use risingwave_common::util::sort_util::{OrderPair, OrderType};
+use risingwave_storage::table::streaming_table::state_table::StateTable;
 use risingwave_storage::StateStore;
 
 use super::error::StreamExecutorResult;
-use super::managed_state::top_n::variants::TOP_N_MAX;
-use super::managed_state::top_n::ManagedTopNState;
+use super::managed_state::top_n::ManagedTopNStateNew;
+use super::top_n::TopNCache;
 use super::top_n_executor::{generate_output, TopNExecutorBase, TopNExecutorWrapper};
 use super::{Executor, ExecutorInfo, PkIndices, PkIndicesRef};
 use crate::executor::top_n::generate_executor_pk_indices_info;
@@ -38,16 +39,12 @@ impl<S: StateStore> AppendOnlyTopNExecutor<S> {
     pub fn new(
         input: Box<dyn Executor>,
         order_pairs: Vec<OrderPair>,
-        offset_and_limit: (usize, Option<usize>),
+        offset_and_limit: (usize, usize),
         pk_indices: PkIndices,
-        store: S,
-        table_id_l: TableId,
-        table_id_h: TableId,
-
-        cache_size: Option<usize>,
-        total_count: (usize, usize),
+        total_count: usize,
         executor_id: u64,
         key_indices: Vec<usize>,
+        state_table: StateTable<S>,
     ) -> StreamExecutorResult<Self> {
         let info = input.info();
         let schema = input.schema().clone();
@@ -60,13 +57,10 @@ impl<S: StateStore> AppendOnlyTopNExecutor<S> {
                 order_pairs,
                 offset_and_limit,
                 pk_indices,
-                store,
-                table_id_l,
-                table_id_h,
-                cache_size,
                 total_count,
                 executor_id,
                 key_indices,
+                state_table,
             )?,
         })
     }
@@ -78,12 +72,6 @@ pub struct InnerAppendOnlyTopNExecutor<S: StateStore> {
     /// Schema of the executor.
     schema: Schema,
 
-    /// `LIMIT XXX`. `None` means no limit.
-    limit: Option<usize>,
-
-    /// `OFFSET XXX`. `0` means no offset.
-    offset: usize,
-
     /// The primary key indices of the `AppendOnlyTopNExecutor`
     pk_indices: PkIndices,
 
@@ -93,17 +81,11 @@ pub struct InnerAppendOnlyTopNExecutor<S: StateStore> {
     /// The order of internal keys of the `TopNExecutor`
     internal_key_order_types: Vec<OrderType>,
 
-    /// We are only interested in which element is in the range of `[offset, offset+limit)`(right
-    /// open interval) but not the rank of such element
-    ///
-    /// We keep two ordered sets. One set stores the elements in the range of `[0, offset)`, and
-    /// another set stores the elements in the range of `[offset, offset+limit)`.
-    managed_lower_state: ManagedTopNState<S, TOP_N_MAX>,
-    managed_higher_state: ManagedTopNState<S, TOP_N_MAX>,
+    /// We are interested in which element is in the range of [offset, offset+limit).
+    managed_state: ManagedTopNStateNew<S>,
 
-    /// Marks whether this is first-time execution. If yes, we need to fill in the cache from
-    /// storage.
-    first_execution: bool,
+    /// In-memory cache of top (N + N * `TOPN_CACHE_HIGH_CAPACITY_FACTOR`) rows
+    cache: TopNCache,
 
     #[expect(dead_code)]
     /// Indices of the columns on which key distribution depends.
@@ -116,26 +98,24 @@ impl<S: StateStore> InnerAppendOnlyTopNExecutor<S> {
         input_info: ExecutorInfo,
         schema: Schema,
         order_pairs: Vec<OrderPair>,
-        offset_and_limit: (usize, Option<usize>),
+        offset_and_limit: (usize, usize),
         pk_indices: PkIndices,
-        store: S,
-        table_id_l: TableId,
-        table_id_h: TableId,
-        cache_size: Option<usize>,
-        total_count: (usize, usize),
+        total_count: usize,
         executor_id: u64,
         key_indices: Vec<usize>,
+        state_table: StateTable<S>,
     ) -> StreamExecutorResult<Self> {
         let (internal_key_indices, internal_key_data_types, internal_key_order_types) =
             generate_executor_pk_indices_info(&order_pairs, &pk_indices, &schema);
 
-        let row_data_types = schema
-            .fields
-            .iter()
-            .map(|field| field.data_type.clone())
-            .collect::<Vec<_>>();
         let ordered_row_deserializer =
             OrderedRowDeserializer::new(internal_key_data_types, internal_key_order_types.clone());
+
+        let num_offset = offset_and_limit.0;
+        let num_limit = offset_and_limit.1;
+        let managed_state =
+            ManagedTopNStateNew::<S>::new(total_count, state_table, ordered_row_deserializer);
+
         Ok(Self {
             info: ExecutorInfo {
                 schema: input_info.schema,
@@ -143,30 +123,11 @@ impl<S: StateStore> InnerAppendOnlyTopNExecutor<S> {
                 identity: format!("AppendOnlyTopNExecutor {:X}", executor_id),
             },
             schema,
-            offset: offset_and_limit.0,
-            limit: offset_and_limit.1,
-            managed_lower_state: ManagedTopNState::<S, TOP_N_MAX>::new(
-                cache_size,
-                total_count.0,
-                store.clone(),
-                table_id_l,
-                row_data_types.clone(),
-                ordered_row_deserializer.clone(),
-                internal_key_indices.clone(),
-            ),
-            managed_higher_state: ManagedTopNState::<S, TOP_N_MAX>::new(
-                cache_size,
-                total_count.1,
-                store,
-                table_id_h,
-                row_data_types,
-                ordered_row_deserializer,
-                internal_key_indices.clone(),
-            ),
+            managed_state,
             pk_indices,
             internal_key_indices,
             internal_key_order_types,
-            first_execution: true,
+            cache: TopNCache::new(num_offset, num_limit),
             key_indices,
         })
     }
@@ -179,86 +140,37 @@ impl<S: StateStore> TopNExecutorBase for InnerAppendOnlyTopNExecutor<S> {
         chunk: StreamChunk,
         epoch: u64,
     ) -> StreamExecutorResult<StreamChunk> {
-        if self.first_execution {
-            self.managed_lower_state.fill_in_cache(epoch).await?;
-            self.managed_higher_state.fill_in_cache(epoch).await?;
-            self.first_execution = false;
-        }
+        let mut res_ops = Vec::with_capacity(self.cache.limit);
+        let mut res_rows = Vec::with_capacity(self.cache.limit);
+        // We have banned streaming query that doesn't have LIMIT, so it is safe to unwrap here
 
-        let num_need_to_keep = self.limit.unwrap_or(usize::MAX);
-        let mut new_ops = vec![];
-        let mut new_rows = vec![];
-
-        for (op, row_ref) in chunk.rows() {
-            assert_eq!(op, Op::Insert);
-
+        // apply the chunk to state table
+        for (_, row_ref) in chunk.rows() {
             let pk_row = row_ref.row_by_indices(&self.internal_key_indices);
             let ordered_pk_row = OrderedRow::new(pk_row, &self.internal_key_order_types);
             let row = row_ref.to_owned_row();
+            self.managed_state
+                .insert(ordered_pk_row.clone(), row.clone(), epoch)?;
 
-            if self.managed_lower_state.total_count() < self.offset {
-                // `elem` is in the range of `[0, offset)`,
-                // we ignored it for now as it is not in the result set.
-                self.managed_lower_state
-                    .insert(ordered_pk_row, row, epoch)?;
-                continue;
-            }
-
-            // We remark that when offset is 0, every input row has nothing to do with
-            // `managed_lower_state`.
-            let element_to_compare_with_upper = if self.offset > 0
-                && &ordered_pk_row < self.managed_lower_state.top_element().unwrap().0
-            {
-                // If the new element is smaller than the largest element in [0, offset),
-                // the largest element may need to move to [offset, offset+limit).
-                let res = self
-                    .managed_lower_state
-                    .pop_top_element(epoch)
-                    .await?
-                    .unwrap();
-                self.managed_lower_state
-                    .insert(ordered_pk_row, row, epoch)?;
-                res
-            } else {
-                (ordered_pk_row, row)
-            };
-
-            if self.managed_higher_state.total_count() < num_need_to_keep {
-                self.managed_higher_state.insert(
-                    element_to_compare_with_upper.0,
-                    element_to_compare_with_upper.1.clone(),
+            self.cache
+                .update(
+                    None,
+                    &mut self.managed_state,
+                    Op::Insert,
+                    ordered_pk_row,
+                    row,
                     epoch,
-                )?;
-                new_ops.push(Op::Insert);
-                new_rows.push(element_to_compare_with_upper.1);
-            } else if self.managed_higher_state.top_element().unwrap().0
-                > &element_to_compare_with_upper.0
-            {
-                let element_to_pop = self
-                    .managed_higher_state
-                    .pop_top_element(epoch)
-                    .await?
-                    .unwrap();
-                new_ops.push(Op::Delete);
-                new_rows.push(element_to_pop.1);
-                new_ops.push(Op::Insert);
-                new_rows.push(element_to_compare_with_upper.1.clone());
-                self.managed_higher_state.insert(
-                    element_to_compare_with_upper.0,
-                    element_to_compare_with_upper.1,
-                    epoch,
-                )?;
-            }
-            // The "else" case can only be that `element_to_compare_with_upper` is larger than
-            // the largest element in [offset, offset+limit), which is already full.
-            // Therefore, nothing happens.
+                    &mut res_ops,
+                    &mut res_rows,
+                )
+                .await?
         }
-        generate_output(new_rows, new_ops, &self.schema)
+        // compare the those two ranges and emit the differantial result
+        generate_output(res_rows, res_ops, &self.schema)
     }
 
     async fn flush_data(&mut self, epoch: u64) -> StreamExecutorResult<()> {
-        self.managed_higher_state.flush(epoch).await?;
-        self.managed_lower_state.flush(epoch).await
+        self.managed_state.flush(epoch).await
     }
 
     fn schema(&self) -> &Schema {
@@ -273,8 +185,10 @@ impl<S: StateStore> TopNExecutorBase for InnerAppendOnlyTopNExecutor<S> {
         &self.info.identity
     }
 
-    async fn init(&mut self, _epoch: u64) -> StreamExecutorResult<()> {
-        Ok(())
+    async fn init(&mut self, epoch: u64) -> StreamExecutorResult<()> {
+        self.managed_state
+            .init_topn_cache(None, &mut self.cache, epoch)
+            .await
     }
 }
 
@@ -285,11 +199,11 @@ mod tests {
     use futures::StreamExt;
     use risingwave_common::array::stream_chunk::StreamChunkTestExt;
     use risingwave_common::array::StreamChunk;
-    use risingwave_common::catalog::{Field, Schema, TableId};
+    use risingwave_common::catalog::{Field, Schema};
     use risingwave_common::types::DataType;
     use risingwave_common::util::sort_util::{OrderPair, OrderType};
-    use risingwave_storage::memory::MemoryStateStore;
 
+    use crate::executor::test_utils::top_n_executor::create_in_memory_state_table;
     use crate::executor::test_utils::MockSource;
     use crate::executor::top_n_appendonly::AppendOnlyTopNExecutor;
     use crate::executor::{Barrier, Epoch, Executor, Message, PkIndices};
@@ -364,98 +278,25 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_append_only_top_n_executor_with_offset() {
-        let order_pairs = create_order_pairs();
-        let source = create_source();
-
-        let top_n_executor = Box::new(
-            AppendOnlyTopNExecutor::new(
-                source as Box<dyn Executor>,
-                order_pairs,
-                (3, None),
-                vec![0, 1],
-                MemoryStateStore::new(),
-                TableId::from(0x2333),
-                TableId::from(0x2334),
-                Some(2),
-                (0, 0),
-                1,
-                vec![],
-            )
-            .unwrap(),
-        );
-        let mut top_n_executor = top_n_executor.execute();
-
-        // consume the init epoch
-        top_n_executor.next().await.unwrap().unwrap();
-        let res = top_n_executor.next().await.unwrap().unwrap();
-        assert_eq!(
-            *res.as_chunk().unwrap(),
-            StreamChunk::from_pretty(
-                "  I I
-                + 10 3
-                +  9 4
-                +  8 5"
-            )
-        );
-        // We added (1, 2, 3, 10, 9, 8).
-        // Now (1, 2, 3) -> (8, 9, 10)
-        // Barrier
-        assert_matches!(
-            top_n_executor.next().await.unwrap().unwrap(),
-            Message::Barrier(_)
-        );
-        let res = top_n_executor.next().await.unwrap().unwrap();
-        assert_eq!(
-            *res.as_chunk().unwrap(),
-            StreamChunk::from_pretty(
-                " I I
-                + 7 6
-                + 3 7
-                + 3 2
-                + 9 9"
-            )
-        );
-        // We added (7, 3, 1, 9).
-        // Now (1, 1, 2) -> (3, 3, 7, 8, 9, 10)
-        // Barrier
-        assert_matches!(
-            top_n_executor.next().await.unwrap().unwrap(),
-            Message::Barrier(_)
-        );
-        let res = top_n_executor.next().await.unwrap().unwrap();
-        assert_eq!(
-            *res.as_chunk().unwrap(),
-            StreamChunk::from_pretty(
-                " I  I
-                + 2  1
-                + 1 13
-                + 2 14
-                + 3 15"
-            )
-        );
-        // We added (1, 1, 2, 3).
-        // Now (1, 1, 1) -> (1, 2, 2, 3, 3, 3, 7, 8, 9, 10)
-    }
-
-    #[tokio::test]
     async fn test_append_only_top_n_executor_with_limit() {
         let order_pairs = create_order_pairs();
         let source = create_source();
+        let state_table = create_in_memory_state_table(
+            &[DataType::Int64, DataType::Int64],
+            &[OrderType::Ascending, OrderType::Ascending],
+            &[0, 1],
+        );
 
         let top_n_executor = Box::new(
             AppendOnlyTopNExecutor::new(
                 source as Box<dyn Executor>,
                 order_pairs,
-                (0, Some(5)),
+                (0, 5),
                 vec![0, 1],
-                MemoryStateStore::new(),
-                TableId::from(0x2333),
-                TableId::from(0x2334),
-                Some(2),
-                (0, 0),
+                0,
                 1,
                 vec![],
+                state_table,
             )
             .unwrap(),
         );
@@ -523,20 +364,22 @@ mod tests {
     async fn test_append_only_top_n_executor_with_offset_and_limit() {
         let order_pairs = create_order_pairs();
         let source = create_source();
+        let state_table = create_in_memory_state_table(
+            &[DataType::Int64, DataType::Int64],
+            &[OrderType::Ascending, OrderType::Ascending],
+            &[0, 1],
+        );
 
         let top_n_executor = Box::new(
             AppendOnlyTopNExecutor::new(
                 source as Box<dyn Executor>,
                 order_pairs,
-                (3, Some(4)),
+                (3, 4),
                 vec![0, 1],
-                MemoryStateStore::new(),
-                TableId::from(0x2333),
-                TableId::from(0x2334),
-                Some(2),
-                (0, 0),
+                0,
                 1,
                 vec![],
+                state_table,
             )
             .unwrap(),
         );
